Create ExploratoreDataAnalysis.r
Exploratory data analysis (EDA) is an iterative process, you will use transformation and visualization to generate questions about your data, search for answers and refine the process for a new question. The idea here is that you let yourself explore the data in a way that allows you to get to the story that it is trying to tell. It will help you frame projects and provide interesting insights to check your intuition and maybe even some of the corporate policies in your organization. Think of EDA as a research and development process where you open your mind to what is possible to find in this space.